refactor(storage): deduplicate SKESK patching and bench S2K code
Extract the repeated filterByTag→s2k.produceKey= loop into
patchSkeskPackets() in streaming-s2k.ts and call it from both the
worker script and the PgpEncryptingStorage direct fallback.

Replace the bench's standalone streamingS2kProduceKey copy with a thin
wrapper around the imported streamingIteratedS2kProduceKey, so the
benchmark exercises the production implementation.
