+++ apps/backend/ee/pgp-s2k-worker/script.ts
@@ -1,6 +1,6 @@
 import { parentPort } from 'worker_threads'
 import * as openpgp from 'openpgp'
-import { streamingIteratedS2kProduceKey } from './streaming-s2k'
+import { patchSkeskPackets } from './streaming-s2k'
 
 if (!parentPort) throw new Error('pgp-s2k worker script must run inside a Worker thread')
 
@@ -24,18 +24,7 @@ parentPort.on('message', async (msg: Request) => {
     })
     const message = await openpgp.readMessage({ binaryMessage: stream })
 
-    // Replace the GenericS2K produceKey on every iterated SKESK packet with
-    // the streaming implementation, which hashes the repeated salt||passphrase
-    // block in 64 KB chunks instead of materialising up to ~65 MB at once.
-    // All other openpgp machinery (CFB session-key unwrap, AEAD, etc.) is unchanged.
-    for (const pkt of message.packets.filterByTag(openpgp.enums.packet.symEncryptedSessionKey)) {
-      const s2k = (pkt as any).s2k
-      if (s2k?.type === 'iterated') {
-        const { algorithm, salt, c } = s2k
-        s2k.produceKey = (passphrase: string, keySizeBytes: number): Promise<Uint8Array> =>
-          Promise.resolve(streamingIteratedS2kProduceKey(algorithm, salt, c, passphrase, keySizeBytes))
-      }
-    }
+    patchSkeskPackets(message)
 
     const [sk] = await openpgp.decryptSessionKeys({ message, passwords: [msg.passphrase] })
     parentPort!.postMessage({
